% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{complexity}
\alias{complexity}
\alias{complexity.default}
\alias{complexity.formula}
\title{Simulated Complexity Measures}
\usage{
complexity(...)

\method{complexity}{default}(x, y, features = "all", ...)

\method{complexity}{formula}(formula, data, features = "all", ...)
}
\arguments{
\item{...}{Further arguments passed to the summarization functions.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{features}{A list of features names or \code{"all"} to include all them.}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class.
The details section describes the valid values for this group.}
}
\value{
A list named by the requested meta-features.
}
\description{
The complexity measures are interested to quantify the the ambiguity of the 
classes, the sparsity and dimensionality of the data and the complexity of
the boundary separating the classes.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"F2"}{}
   \item{"F3"}{}
   \item{"F4"}{}
   \item{"N1"}{}
   \item{"N3"}{}
   \item{"N4"}{}
   \item{"T1"}{}
   \item{"LSC"}{}
   \item{"L2"}{}
   \item{"L3"}{}
   \item{"Density"}{}
   \item{"ClsCoef"}{}
   \item{"Hubs"}{}
 }
}
\examples{
## Extract all complexity measures using formula
complexity(Species ~ ., iris)

## Extract some complexity measures
complexity(iris[1:4], iris[5], c("F2", "F3", "F4"))
}
\references{
Tin K Ho and Mitra Basu. Complexity measures of supervised classification 
 problems. IEEE Transactions on Pattern Analysis and Machine Intelligence, 
 24, 3, 289--300, 2002.

 Albert Orriols-Puig, Nuria Macia and Tin K Ho. Documentation for the data 
 complexity library in C++. Technical Report. La Salle - Universitat Ramon 
 Llull, 2010.
}
\concept{meta-features}
