% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity.R
\name{ls.complexity}
\alias{ls.complexity}
\title{List the simulated complexity measures}
\usage{
ls.complexity()
}
\value{
A list of the simulated complexity measures names.
}
\description{
List the simulated complexity measures
}
\examples{
ls.complexity()
}
