% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.R
\name{ls.simulated}
\alias{ls.simulated}
\title{List the simulated simulated measures}
\usage{
ls.simulated()
}
\value{
A list of the simulated simulated measures names.
}
\description{
List the simulated simulated measures
}
\examples{
ls.simulated()
}
