% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated.R
\name{simulated}
\alias{simulated}
\alias{simulated.default}
\alias{simulated.formula}
\title{Simulated Complexity Measures}
\usage{
simulated(...)

\method{simulated}{default}(x, y, features = "all", ...)

\method{simulated}{formula}(formula, data, features = "all", ...)
}
\arguments{
\item{...}{Further arguments passed to the summarization functions.}

\item{x}{A data.frame contained only the input attributes.}

\item{y}{A factor response vector with one label for each row/component of x.}

\item{features}{A list of features names or \code{"all"} to include all them.}

\item{formula}{A formula to define the class column.}

\item{data}{A data.frame dataset contained the input attributes and class.
The details section describes the valid values for this group.}
}
\value{
A list named by the requested meta-features.
}
\description{
The complexity measures are interested to quantify the the ambiguity of the 
classes, the sparsity and dimensionality of the data and the complexity of
the boundary separating the classes.
}
\details{
The following features are allowed for this method:
 \describe{
   \item{"F2"}{}
   \item{"F3"}{}
   \item{"F4"}{}
   \item{"N1"}{}
   \item{"N3"}{}
   \item{"N4"}{}
   \item{"T1"}{}
   \item{"LSC"}{}
   \item{"L2"}{}
   \item{"L3"}{}
   \item{"Density"}{}
   \item{"ClsCoef"}{}
   \item{"Hubs"}{}
 }
}
\examples{
## Extract all complexity measures using formula
simulated(Species ~ ., iris)

## Extract some complexity measures
simulated(iris[1:4], iris[5], c("F2", "F3", "F4"))
}
\references{
Ana C. Lorena, Luis P. F. Garcia, Jens  Lehmann, Marcilio C. P. de Souto and
 Tin k. Ho. How Complex is your classification problem? A survey on measuring
 classification complexity. arXiv:1808.03591, 2018.
}
\concept{meta-features}
